Profile: Code cleanup and project restructure
