+++ game/src/game/char.c
@@ -41,7 +41,7 @@
 #include "game/egoboo.h"
 #include "game/module/PassageHandler.hpp"
 #include "game/audio/AudioSystem.hpp"
-#include "game/ProfileSystem.hpp"
+#include "game/profiles/ProfileSystem.hpp"
 
 #include "game/ChrList.h"
 #include "game/EncList.h"
