+++ game/src/game/egoboo.c
@@ -38,8 +38,8 @@
 #include "game/particle.h"
 #include "game/enchant.h"
 #include "game/collision.h"
-#include "game/Profile.hpp"
-#include "game/ProfileSystem.hpp"
+#include "game/profiles/Profile.hpp"
+#include "game/profiles/ProfileSystem.hpp"
 
 #include "game/ChrList.h"
 #include "game/EncList.h"
