+++ game/src/game/game.c
@@ -45,12 +45,12 @@
 #include "game/module/PassageHandler.hpp"
 #include "game/graphics/CameraSystem.hpp"
 #include "game/audio/AudioSystem.hpp"
-#include "game/ProfileSystem.hpp"
+#include "game/profiles/ProfileSystem.hpp"
 
 #include "game/char.h"
 #include "game/particle.h"
 #include "game/enchant.h"
-#include "game/Profile.hpp"
+#include "game/profiles/Profile.hpp"
 #include "game/mesh.h"
 #include "game/physics.h"
