+++ game/src/game/profiles/RandomName.cpp
@@ -20,7 +20,7 @@
 /// @brief This class is used for random name generation
 /// @author Johan Jansen
 
-#include "game/RandomName.hpp"
+#include "game/profiles/RandomName.hpp"
 #include "egolib/strutil.h"
 #include "egolib/fileutil.h"
 #include "egolib/math/Random.hpp"
