+++ game/src/game/script_compile.h
@@ -23,7 +23,7 @@
 
 #include "game/egoboo_typedef.h"
 #include "game/script.h"
-#include "game/Profile.hpp"
+#include "game/profiles/Profile.hpp"
 
 //--------------------------------------------------------------------------------------------
 //--------------------------------------------------------------------------------------------
